+++ datafusion/core/Cargo.toml
@@ -110,8 +110,8 @@ doc-comment = "0.3"
 env_logger = "0.10"
 parquet-test-utils = { path = "../../parquet-test-utils" }
 rstest = "0.16.0"
-test-utils = { path = "../../test-utils" }
 sqlparser = "0.27"
+test-utils = { path = "../../test-utils" }
 
 [[bench]]
 harness = false
